Refactor Document class
